New App w/ Redux - Duck Pattern
