--- src/store/actions/favorites.js
@@ -1,16 +1,0 @@
-export const addFavoriteRequest = repository => ({
-  type: 'ADD_FAVORITE_REQUEST',
-  payload: { repository },
-});
-
-// REQUEST -> SAGA -> CHAMADA API -> SUCCESS
-
-export const addFavoriteSuccess = data => ({
-  type: 'ADD_FAVORITE_SUCCESS',
-  payload: { data },
-});
-
-export const addFavoriteFailure = error => ({
-  type: 'ADD_FAVORITE_FAILURE',
-  payload: { error },
-});
