+++ src/store/sagas/index.js
@@ -1,7 +1,9 @@
 import { all, takeLatest } from 'redux-saga/effects';
 
+import { Types as FavoriteTypes } from '../ducks/favorites';
+
 import { addFavorite } from './favorites';
 
 export default function* rootSaga() {
-  yield all([takeLatest('ADD_FAVORITE_REQUEST', addFavorite)]);
+  yield all([takeLatest(FavoriteTypes.ADD_REQUEST, addFavorite)]);
 }
